refactor: decompose psa_gui into ui submodule in psa_package (fixes #1970)
